devops testing finished

+++ database/migrations/2022_05_22_074230_create_dashboards_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('year')->unsigned();
             $table->integer('quartile')->unsigned();
             $table->text('course');
-            $table->integer('EC')->unsigned();
+            $table->decimal('EC')->unsigned();
             $table->timestamps();
         });
     }
